+++ internal/affine/affine.go
@@ -25,11 +25,11 @@ func add(lhs, rhs []float64, dim int) []float64 {
 	return result
 }
 
-func mulSquare(lhs, rhs []float64, dim int) []float64 {
-	result := make([]float64, len(lhs))
+func mulSquare(lhs, rhs []float32, dim int) []float32 {
+	result := make([]float32, len(lhs))
 	for i := 0; i < dim; i++ {
 		for j := 0; j < dim; j++ {
-			e := 0.0
+			e := float32(0.0)
 			for k := 0; k < dim; k++ {
 				e += lhs[i*dim+k] * rhs[k*dim+j]
 			}
